Move autocmds and add another treesitter lang
Moved the autocmds lua file to actually be in the lua directory. I
forgot to put it there when I created it. Additionally, made sure that
treesitter installs the help language. I imagine I'll be doing this for
all kinds of languages. I'll try and refine my disable function so it's
more flexible.